use separate zod schema for sortBy queries
