+++ packages/api/src/schema/zod/schemas/withTvlQuery.ts
@@ -1,62 +1,10 @@
 import z from '..'
 
-export const WithTvlQuerySchema = z
-	.object({
-		withTvl: z
-			.enum(['true', 'false'])
-			.default('false')
-			.describe('Toggle whether the route should calculate the TVL from shares')
-			.transform((val) => val === 'true')
-			.openapi({ example: 'false' }),
-		sortByTvl: z
-			.enum(['asc', 'desc'])
-			.optional()
-			.describe('Sort results in asc or desc order of TVL value')
-			.openapi({ example: 'desc' }),
-		sortByTotalStakers: z
-			.enum(['asc', 'desc'])
-			.optional()
-			.describe('Sort results in asc or desc order of total stakers')
-			.openapi({ example: 'desc' }),
-		sortByTotalAvs: z
-			.enum(['asc', 'desc'])
-			.optional()
-			.describe(
-				'Sort results in asc or desc order of total AVS (only valid for Operator queries)'
-			)
-			.openapi({ example: 'desc' }),
-		sortByTotalOperators: z
-			.enum(['asc', 'desc'])
-			.optional()
-			.describe(
-				'Sort results in asc or desc order of total AVS (only valid for AVS queries)'
-			)
-			.openapi({ example: 'desc' }),
-		sortOperatorsByTvl: z
-			.enum(['asc', 'desc'])
-			.optional()
-			.describe(
-				'Sort Operators of a given AVS in asc or desc order of TVL (only valid for individual AVS queries)'
-			)
-			.openapi({ example: 'desc' })
-	})
-	.refine(
-		(data) => {
-			const sortByFields = [
-				data.sortByTvl,
-				data.sortByTotalStakers,
-				data.sortByTotalAvs,
-				data.sortByTotalOperators
-			].filter((field) => field !== undefined)
-			return sortByFields.length <= 1
-		},
-		{
-			message: 'Only one sortBy option can be used',
-			path: [
-				'sortByTvl',
-				'sortByTotalStakers',
-				'sortByTotalAvs',
-				'sortByTotalOperators'
-			]
-		}
-	)
+export const WithTvlQuerySchema = z.object({
+	withTvl: z
+		.enum(['true', 'false'])
+		.default('false')
+		.describe('Toggle whether the route should calculate the TVL from shares')
+		.transform((val) => val === 'true')
+		.openapi({ example: 'false' })
+})
